@@ -35,19 +35,18 @@
   },
   "main": "public/electron.js",
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.2.5",
-    "@mui/material": "^5.2.7",
-    "crc": "^4.1.0",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.6.2",
+    "@mui/material": "^5.6.4",
+    "crc": "^4.1.1",
     "cross-env": "^7.0.3",
     "electron-is-dev": "^2.0.0",
-    "react": "^17.0.2",
+    "react": "^18.1.0",
     "react-contexify": "^5.0.0",
-    "react-dom": "^17.0.2",
-    "react-router": "^6.2.1",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0"
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "react-start": "react-scripts start",
@@ -72,16 +71,16 @@
     ]
   },
   "devDependencies": {
-    "concurrently": "^7.0.0",
-    "electron": "^16.0.6",
-    "electron-builder": "^22.14.5",
-    "eslint": "^8.6.0",
+    "concurrently": "^7.1.0",
+    "electron": "^18.2.0",
+    "electron-builder": "^23.0.3",
+    "eslint": "^8.15.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "wait-on": "^6.0.0"
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.5.0",
+    "wait-on": "^6.0.1"
   },
-  "packageManager": "yarn@3.1.1"
+  "packageManager": "yarn@3.2.0"
 }
