@@ -1,18 +1,19 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './components/App';
 import * as serviceWorker from './serviceWorker';
 import MainContextProvider from './contexts/MainContextProvider';
 import CryptoContextProvider from './contexts/CryptoContextReducer';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
   <MainContextProvider>
     <CryptoContextProvider>
       <App />
     </CryptoContextProvider>
   </MainContextProvider>,
-  document.getElementById('root'),
 );
 
 // If you want your app to work offline and load faster, you can change
